fixup: Re-enables user to set an arbitrary password when protecting a public  authenticated share.
I initially thought it was necessary to prevent users from setting an arbitrary password in such case (except when the password is given via Talk).
But, eventually, I've figured out it was not necessary to drop support for that scenario.

This commit is part of https://github.com/nextcloud/server/issues/31005

Signed-off-by: Cyrille Bollu <cyrpub@bollu.be>